refactor(next): move nextServerAdapter to adapters/ directory
Co-locates the Next ServerAdapter impl with the RouterAdapter impl under
packages/next/src/adapters/. Both are framework bridges, not utilities.
